+++ Playgrounds/JSONSynchronization.playground/Contents.swift
@@ -47,7 +47,7 @@ class Person : Record {
         super.init(row)
     }
     
-    override var persistedDictionary: [String: DatabaseValueConvertible?] {
+    override var persistentDictionary: [String: DatabaseValueConvertible?] {
         return ["id": id, "name": name]
     }