+++ Playgrounds/Record.playground/Contents.swift
@@ -49,7 +49,7 @@ class Person : Record {
         super.init(row)
     }
     
-    override var persistedDictionary: [String: DatabaseValueConvertible?] {
+    override var persistentDictionary: [String: DatabaseValueConvertible?] {
         return ["id": id, "firstName": firstName, "lastName": lastName]
     }